+++ skeleton/settings_qa_smart_site.py
@@ -1,0 +1,8 @@
+from foundry import settings as foundry_settings
+
+from skeleton.settings_qa_base import *
+
+
+FOUNDRY['layers'] = ('smart', 'mid', 'basic',)
+
+foundry_settings.compute_settings(sys.modules[__name__])